Remove the code related to the JavaScript browser GUI.
As part of an effort to focus the code to be more easily managed and
completed, the front end code will be removed entirely. It will be
very out of date by the time work on the GUI resumes. The todo is also
quite old at this point.

The design philosophy will change from trying to complete Federation
all at once to starting with a small, unambitious game engine and
iterating increasing complexity from there. That way, there will be a
playable game at almost every stage of development.

Federation itself will now consist of multiple modular engines that
work together to meet all of the planned features. Smaller games
focusing on certain subsets will be created along the way so that the
code is well-tested before the modules are united into Federation.